requirements.txt
streamlit
flask
flask-cors
python-dotenv
openai
vapi
requests